requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
antlr4-python3-runtime
anyio
appnope
asttokens
async-timeout
attrs
backoff
beautifulsoup4
cachetools
certifi
cffi
chardet
charset-normalizer
click
cohere
coloredlogs
comm
contourpy
cryptography
cycler
dataclasses-json
debugpy
decorator
Deprecated
dirtyjson
distro
effdet
emoji
eval_type_backport
exceptiongroup
executing
fastapi
fastavro
filelock
filetype
flatbuffers
fonttools
frozenlist
fsspec
google-api-core
google-auth
google-cloud-vision
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
html5lib
httpcore
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
iopath
ipykernel
ipython
ipywidgets
jedi
Jinja2
jiter
joblib
jsonpath-python
jupyter_client
jupyter_core
jupyterlab_widgets
kiwisolver
langdetect
layoutparser
llama-index-agent-openai
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-postprocessor-cohere-rerank
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
lxml
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mpmath
multidict
mypy-extensions
ndjson
nest-asyncio
networkx
nltk
numpy
olefile
omegaconf
onnx
onnxruntime
openai
opencv-python
packaging
pandas
parameterized
parso
pdf2image
pdfminer.six
pdfplumber
pexpect
pi_heif
pikepdf
pillow
platformdirs
portalocker
prompt_toolkit
propcache
proto-plus
protobuf
psutil
ptyprocess
pure_eval
pyasn1
pyasn1_modules
pycocotools
pycparser
pydantic>=2.0.0
pydantic_core
Pygments
pyparsing
pypdf
pypdfium2
python-dateutil
python-iso639
python-magic
python-multipart
python-oxmsg
pytz
PyYAML
pyzmq
RapidFuzz
regex
requests
requests-toolbelt
rsa
safetensors
scipy
six
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
striprtf
sympy
tenacity
tiktoken
timm
tokenizers
torch
torchvision
tornado
tqdm
traitlets
transformers
types-requests
typing-inspect
typing_extensions
tzdata
unstructured
unstructured-client
unstructured-inference
unstructured.pytesseract
urllib3
uvicorn
wcwidth
webencodings
widgetsnbextension
wrapt
yarl
pydantic-settings
sqlalchemy>=1.4.0